refactor: centralize ModuleConfig interface
