+++ src/modules/module-types.ts
@@ -1,0 +1,9 @@
+import type { ComponentType } from 'react';
+
+export interface ModuleConfig {
+    id: string;
+    name: string;
+    dependencies: string[];
+    apiBasePath: string;
+    UiComponent: ComponentType<unknown>;
+}
